# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MASGuardEval API",
    "description": "Golden-dataset-based evaluation and observability framework for multi-agent LLM systems.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/scenarios": {
      "get": {
        "summary": "Scenarios",
        "operationId": "scenarios_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Scenarios Scenarios Get"
                }
              }
            }
          }
        }
      }
    },
    "/evaluate/{scenario_id}": {
      "get": {
        "summary": "Evaluate",
        "operationId": "evaluate_evaluate__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evaluate Evaluate  Scenario Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard": {
      "get": {
        "summary": "Dashboard",
        "operationId": "dashboard_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Dashboard Dashboard Get"
                }
              }
            }
          }
        }
      }
    },
    "/experiments": {
      "get": {
        "summary": "Experiments",
        "operationId": "experiments_experiments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Experiments Experiments Get"
                }
              }
            }
          }
        }
      }
    },
    "/propagation/{scenario_id}": {
      "get": {
        "summary": "Propagation",
        "operationId": "propagation_propagation__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Propagation Propagation  Scenario Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scalability/batch": {
      "get": {
        "summary": "Scalability Batch",
        "operationId": "scalability_batch_scalability_batch_get",
        "parameters": [
          {
            "name": "max_workers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 4,
              "title": "Max Workers"
            }
          },
          {
            "name": "chunk_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Chunk Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Scalability Batch Scalability Batch Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluator/agreement": {
      "post": {
        "summary": "Evaluator Agreement",
        "operationId": "evaluator_agreement_evaluator_agreement_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgreementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Evaluator Agreement Evaluator Agreement Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/project-docs": {
      "get": {
        "summary": "Project Docs Home",
        "operationId": "project_docs_home_project_docs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/project-docs/{doc_name}": {
      "get": {
        "summary": "Project Doc",
        "operationId": "project_doc_project_docs__doc_name__get",
        "parameters": [
          {
            "name": "doc_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgreementRequest": {
        "properties": {
          "labels_a": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Labels A"
          },
          "labels_b": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Labels B"
          }
        },
        "type": "object",
        "required": [
          "labels_a",
          "labels_b"
        ],
        "title": "AgreementRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}